move extension file outside of /agent_builder
