+++ front/components/agent_builder/instructions/AgentBuilderInstructionsEditor.tsx
@@ -17,7 +17,7 @@ import { InstructionTipsPopover } from "@app/components/agent_builder/instructio
 import { useBlockInsertDropdown } from "@app/components/agent_builder/instructions/useBlockInsertDropdown";
 import { AgentInstructionDiffExtension } from "@app/components/editor/extensions/agent_builder/AgentInstructionDiffExtension";
 import { BlockInsertExtension } from "@app/components/editor/extensions/agent_builder/BlockInsertExtension";
-import { HeadingExtension } from "@app/components/editor/extensions/agent_builder/HeadingExtension";
+import { HeadingExtension } from "@app/components/editor/extensions/HeadingExtension";
 import { InstructionBlockExtension } from "@app/components/editor/extensions/agent_builder/InstructionBlockExtension";
 import { EmojiExtension } from "@app/components/editor/extensions/EmojiExtension";
 import { EmptyLineParagraphExtension } from "@app/components/editor/extensions/EmptyLineParagraphExtension";
