+++ front/components/editor/SkillInstructionsEditor.tsx
@@ -8,7 +8,7 @@ import { StarterKit } from "@tiptap/starter-kit";
 import { useEffect, useMemo, useRef } from "react";
 
 import { AgentInstructionDiffExtension } from "@app/components/editor/extensions/agent_builder/AgentInstructionDiffExtension";
-import { HeadingExtension } from "@app/components/editor/extensions/agent_builder/HeadingExtension";
+import { HeadingExtension } from "@app/components/editor/extensions/HeadingExtension";
 import { ListItemExtension } from "@app/components/editor/extensions/ListItemExtension";
 import { OrderedListExtension } from "@app/components/editor/extensions/OrderedListExtension";
 import type { KnowledgeItem } from "@app/components/editor/extensions/skill_builder/KnowledgeNode";
